Handle ref not starting with refs/
This commit adds better handling and logging for the case
were ref does not start with refs/.... We already had something
in place to handle PR events. Now we make sure to do this only
if we have no info in ref. Otherwise we are logging the event name and the
context for easier debugging if something went wrong.